pyproject.toml
[project]
name = "pfb-imaging"
version = "0.0.6"
description = "Radio interferometric imaging suite based on a preconditioned forward-backward approach"
authors = [
    { name = "landmanbester", email = "lbester@sarao.ac.za" }
]
license = {text = "MIT"}
readme = "README.rst"
requires-python = ">=3.10, <3.13"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Astronomy"
]

dependencies = [
    "typer>=0.12.0",
    "hip-cargo @ git+https://github.com/landmanbester/hip-cargo.git@genfunc",
]

[project.urls]
Homepage = "https://github.com/ratt-ru/pfb-imaging"

[project.optional-dependencies]
# Full scientific stack
full = [
    "katbeam>=0.1",
    "codex-africanus[dask,scipy,astropy,python-casacore]>=0.4.2",
    "dask-ms[s3,xarray,zarr]>=0.2.23",
    "reproject",
    "ray[default]",
    "scikit-image>=0.24.0",
    "PyWavelets>=1.7.0",
    "numexpr>=2.10.1",
    "ducc0>=0.35.0",
    "sympy>=1.9",
    "stimela>=2.0.2",
    "streamjoy>=0.0.8",
    "tbb>=2021.13.1",
    "jax[cpu]>=0.4.31",
    "lz4>=4.3.3",
    "bokeh>=3.1.0",
    "regions>=0.9",
    "psutil>=5.9.8",
    "matplotlib>=3.9.2",
]

dev = [
    "pytest>=8.0.0",
    "tbump>=6.11.0",
    "ruff>=0.8.0",
    "pre-commit>=3.0.0",
    "mypy>=1.0.0",
]


[project.scripts]
pfb = "pfb_imaging.cli:app"

[build-system]
requires = ["uv_build>=0.8.3,<0.10.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Tests that take more time to run",
]